% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCIC.R
\name{load.umi.count.from.file}
\alias{load.umi.count.from.file}
\title{load.UMI.count.from.file}
\usage{
load.umi.count.from.file(csv_path)
}
\description{
This function loads the UMI count matrix from the standard .csv file
}
\examples{
load.umi.count.from.file
}
\keyword{file}
\keyword{load}
\keyword{umi}
